Add Main class, scrap_repo_page method to scrapper rename developer class to scraper class
