+++ lib/address.rb
@@ -6,11 +6,22 @@ class AddressUri
     @date = ""
   end
 
-  def geenrate_address
+  def generate_address
     address = "https://github.com/trending"
     if @developer
-      address = address + "/developers"
+      address += "/developers"
     end
+    if @language.size > 0
+      address += "/" + @language
+    end
+    if @date == "week"
+      address += "?since=weekly"
+    elsif @date == "month"
+      address += "?since=monthly"
+    elsif @language.size > 0
+      address += "?since=daily"
+    end
+
     address
   end
